Add Mode related classes, integrate EventHandler
A ModeSequence class was created to encapsulate a hierarchical set of
main and submodes that can be cycled through. The two buttons on the
audiobox directly cycle through each of these modes.

The EventHandler class is now fully integrated. This supports event-driven
operation of tasks via a publish (emit) and subscribe system, and eliminates
the previous task-to-task queues, which had become become "spaghetti"-like.

As an example of the power of the new method, when one task emits an
"EVENT_MODE_CHANGED" event, all other tasks that subscribe to that event
are notified. Previously, each other task would require an individual queue
connection to the publishing task.

The EventHandler currently runs at the same priority level as other task;
however, this may need to be changed in the future to ensure events are
always processed.