Add wifi connection timeout, optimize blending
Wifi connection will timeout after a #defined period (currently 10 sec).
Behavior upon failing to connect has not yet been fully implemented but
this will allow the system to be responsive to the power switch even if
wifi does not connect.

LED blending between modes has been optimized to match the speed at which
the servo moves. This helps prevent the previous long delays in blending
between modes when this blend duration was hardcoded.