Refactor button task
The button task was originally taking care of mode change as well. This
was refactored into a separate mode task that subscribes to button press
events and emits a mode change event when appropriate. Meanwhile, the
button task is simply responsible for advancing the button state machines
and emitting a button press event when appropriate.

This change should simplify emulation of button presses via web or other
sources -- we simply need to emit a button press event with the appropriate
parameters and the mode change task will take care of the rest.