Update wifi status on all web pages
Every html page now has a status section that shows the current WiFi network
and connection status. Buttons with a "needs-wifi" class will be disabled
if the wifi network is not connected.

The goal is that the current set of pages becomes the standard preferences
menu, accessible in setup mode or during operation.