Fixed memory issues with json and web server
The web server and Spotify json deserialization were contending for heap
space and causing both tasks to intermittenly fail. The following changes
have seemingly resolved the issue:
- Reallocation of task stack sizes, based on profiling (see commented
code in task_eventhandler). A lot of stack space was unused and is now
reclaimed.
- Use of ArduinoJson's filter method to reduce space required for
deserialization.
- Change to using StaticJsonDocument instead of Dynamic. This coupled
with the above change reduced heap usage from ~20k to stack usage of ~2k.
Note that we still dynamically allocate up to ~6k of String memory on the
heap to capture the HTTP response.

Additionally, ESP.getFreeHeap() does not appear to provide an accurate
reading of remaining heap space. get_memory_stats() was added as a utility
function to provide much more detail on heap usage.