Web control with album art and Spotify status
Updates made to have the web control automatically update Spotify user
name and status (Active or Inactive) based on events passed from the box.

The server also provides the URL of the album art, and client-side
javascript uses this to determine the "large" (300px) album art URL
and displays that instead.

Other changes were made to make the power switch and reboot operations
more responsive and event-based. This includes adding an interrupt during
setup that triggers on the power switch. This prevents the switch from
becoming unresponsive during the pre-task setup.

Using power switch interrupt during full operation seems to be unstable
so I am resorting to using a polling approach to check the state of the
power switch.