Completed main web control workflows
There are now 3 workflows pertaining to the web control:

1. **Control**: This is the default mode when the device has an existing
WiFi and Spotify account and allows the user to change modes, but does
not allow changes to the WiFi or Spotify credentials.

2. **Setup**: This mode is accessed by the user holding down the top
button during bootup and allows WiFi and Spotify changes to be made.

3. **WiFi Setup**: This mode is automatically entered if the device
cannot connect to a WiFi network. The device creates a WiFi access point
that the user can use to setup the WiFi credentials. Note that this
requires the user to manually connect to 192.168.4.1. Ideally this
would have a captive-portal-like behavior in the future that
immediately sends the user to the setup page upon connection to the AP.