Change max LED brightness to avoid brownout
Over the past few months the audiobox would every so often seem to get
stuck, with the LEDs fixed to a bright image and the servo stuttering.
When in this state, the box would be unresponsive to power switch
toggling, and even unplugging/replugging would sometimes not work.
Plugging into the laptop USB would often restart it.

Using a bench power supply I can see that at 5V, the box consumes ~3A
of current when the LEDs are at full white. This just so happens to be
the max current rating for the power supply I am using. Setting the
bench power supply to a current limit of < 3A, I was able to reproduce
the unresponsive box and stuttering servo behavior.

The fix is to reduce the LED brightness to avoid brownout. The current
setting of 60 (out of 255) leads to a max current of ~1.8A at full
white.